fix(auth): remove server dependencies from apollo-link adapters
